requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.2
distro==1.8.0
dnspython==2.4.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
loguru==0.7.2
numpy==1.26.2
openai==1.3.2
pinecone-client==2.2.4
pydantic==2.5.1
pydantic_core==2.14.3
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
